backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
pandas==2.2.2
pytest==8.3.3
rank-bm25==0.2.2
python-docx==1.1.0
fpdf2==2.7.9
matplotlib==3.8.0
Pillow==10.1.0
elasticsearch==8.15.1
google-generativeai==0.8.2
